[BET-18] Delete vercel handler
