+++ cron/package.json
@@ -6,9 +6,14 @@
   "license": "MIT",
   "devDependencies": {
     "serverless-plugin-typescript": "^2.1.4",
-    "typescript": "^4.9.5"
+    "typescript": "^4.9.5",
+    "@types/yargs": "^17.0.22",
+    "ts-node": "^10.9.1",
+    "yargs": "^17.6.2"
   },
   "dependencies": {
+    "@pythnetwork/pyth-aptos-js": "^1.0.0",
+    "aptos": "^1.6.0",
     "axios": "^1.3.1"
   }
 }
