+++ cron/serverless.yml
@@ -16,7 +16,7 @@ functions:
   rateHandlerMain:
     handler: handler.run
     events:
-      - schedule: cron(0 * * * ? *) # every 0 minute i.e. every hour
+      - schedule: cron(0 0,6,12,18 * * ? *) # every 0 minute i.e. every hour
     environment:
       BETOS_ADDRESS: "0xb7819e2435c024bd2a84f7527e6b752882a38a0f8835542fa893ac5f511d58a4"
       NETWORK: "mainnet"
